Tp 1336 fix repeating activities (#224)
Determining current activity is now driven solely by cache update, and activity can change only when actual status in activity array changes.

In addition to this polling for activities is now done on demand, only when we have no current activity. 
Tested this with subscription disabled and it works fine, also tested that stopPolling does not cause an error if there is no polling active.